feat(vendure-scripts): cleanup and linting
