+++ packages/vendure-scripts/src/assign-all-customers-to-channel.ts
@@ -6,13 +6,33 @@ import {
   TransactionalConnection,
 } from '@vendure/core';
 import { IsNull } from 'typeorm';
-export function getCustomers(
+import { assignEntitiesToChannel } from './util';
+
+export async function assignCustomersToChannel(
+  sourceChannelId: ID,
+  targetChannelId: ID,
+  injector: Injector,
+  ctx: RequestContext,
+  batch: number = 10
+): Promise<void> {
+  await assignEntitiesToChannel<Customer>(
+    sourceChannelId,
+    targetChannelId,
+    injector,
+    ctx,
+    getCustomers,
+    batch
+  );
+}
+
+function getCustomers(
   ctx: RequestContext,
   sourceChanneId: ID,
   injector: Injector,
-  take: number = 10,
-  skip: number = 0
+  skip: number = 0,
+  take: number = 10
 ): Promise<Customer[]> {
+  console.log(`Getting customers ${skip} to ${take + skip}`);
   const conn = injector.get(TransactionalConnection);
   const customerRepo = conn.getRepository(ctx, Customer);
   return customerRepo
