+++ packages/vendure-scripts/src/assign-all-orders-to-channel.ts
@@ -5,13 +5,31 @@ import {
   RequestContext,
   TransactionalConnection,
 } from '@vendure/core';
+import { assignEntitiesToChannel } from './util';
 
-export function getOrders(
+export async function assignOrdersToChannel(
+  sourceChannelId: ID,
+  targetChannelId: ID,
+  injector: Injector,
+  ctx: RequestContext,
+  batch: number = 10
+): Promise<void> {
+  await assignEntitiesToChannel<Order>(
+    sourceChannelId,
+    targetChannelId,
+    injector,
+    ctx,
+    getOrders,
+    batch
+  );
+}
+
+function getOrders(
   ctx: RequestContext,
   sourceChanneId: ID,
   injector: Injector,
-  take: number = 10,
-  skip: number = 0
+  skip: number = 0,
+  take: number = 10
 ): Promise<Order[]> {
   const conn = injector.get(TransactionalConnection);
   const orderRepo = conn.getRepository(ctx, Order);
