+++ packages/vendure-scripts/src/copy-database.ts
@@ -1,3 +1,10 @@
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+// Some eslint rules are disabled, because there are no types available for 'prompt-confirm'
+/* eslint-disable @typescript-eslint/no-var-requires */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+
 const Confirm = require('prompt-confirm');
 import { spawn } from 'child_process';
 
@@ -60,10 +67,10 @@ export async function insertIntoDb(
 async function execute(command: string): Promise<void> {
   return new Promise((resolve, reject) => {
     const proc = spawn(command, [], { stdio: 'inherit', shell: true });
-    proc.on('error', function (error: any) {
+    proc.on('error', function (error) {
       reject(new Error(error.message));
     });
-    proc.on('exit', function (code: any) {
+    proc.on('exit', function (code) {
       if (code !== 0) {
         reject(new Error('exited with code ' + code));
       } else {
