+++ packages/vendure-scripts/src/index.ts
@@ -1,5 +1,4 @@
-export * from './assign-products/assign-all-products-to-channel';
-export * from './assign-products/assign-products-to-channel';
-export * from './assign-orders/assign-all-orders-to-channel';
-export * from './assign-customers/assign-all-customers-to-channel';
+export * from './assign-all-products-to-channel';
+export * from './assign-all-customers-to-channel';
+export * from './assign-all-orders-to-channel';
 export * from './copy-database';
