+++ packages/vendure-scripts/vitest.config.js
